Move SimulatorInstaller commands to main CLI
