+++ XHarness.sln
@@ -26,10 +26,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.XHarness.i
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.XHarness.TestRunners.Tests", "tests\Microsoft.DotNet.XHarness.TestRunners.Tests\Microsoft.DotNet.XHarness.TestRunners.Tests.csproj", "{05AB368F-2B27-426A-B5E9-01F52E297269}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tools", "tools", "{D7D3EC29-8CF4-4DA5-BF41-FE5A51593753}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.XHarness.SimulatorInstaller", "tools\Microsoft.DotNet.XHarness.SimulatorInstaller\Microsoft.DotNet.XHarness.SimulatorInstaller.csproj", "{2C1B5C97-7E17-45D4-8337-B4CBD6125CD7}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.XHarness.Common", "src\Microsoft.DotNet.XHarness.Common\Microsoft.DotNet.XHarness.Common.csproj", "{E6E90648-E24D-4681-B082-776DF3F4B8F6}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.XHarness.Common.Tests", "tests\Microsoft.DotNet.XHarness.Common.Tests\Microsoft.DotNet.XHarness.Common.Tests.csproj", "{EB718364-434C-43F7-AE12-74F272B86C45}"
@@ -76,10 +72,6 @@ Global
 		{05AB368F-2B27-426A-B5E9-01F52E297269}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{05AB368F-2B27-426A-B5E9-01F52E297269}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{05AB368F-2B27-426A-B5E9-01F52E297269}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2C1B5C97-7E17-45D4-8337-B4CBD6125CD7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2C1B5C97-7E17-45D4-8337-B4CBD6125CD7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2C1B5C97-7E17-45D4-8337-B4CBD6125CD7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2C1B5C97-7E17-45D4-8337-B4CBD6125CD7}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E6E90648-E24D-4681-B082-776DF3F4B8F6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E6E90648-E24D-4681-B082-776DF3F4B8F6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E6E90648-E24D-4681-B082-776DF3F4B8F6}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -116,7 +108,6 @@ Global
 		{10A26931-A5C8-4256-B8AC-0984CAAAB93D} = {1018BF4C-BD8F-49D7-9797-758E68B9FC9D}
 		{3406110E-65B2-4786-B723-C05B0FF66131} = {1018BF4C-BD8F-49D7-9797-758E68B9FC9D}
 		{05AB368F-2B27-426A-B5E9-01F52E297269} = {1018BF4C-BD8F-49D7-9797-758E68B9FC9D}
-		{2C1B5C97-7E17-45D4-8337-B4CBD6125CD7} = {D7D3EC29-8CF4-4DA5-BF41-FE5A51593753}
 		{E6E90648-E24D-4681-B082-776DF3F4B8F6} = {713ACDCF-301D-4C96-8B75-5BF2A35A326D}
 		{EB718364-434C-43F7-AE12-74F272B86C45} = {1018BF4C-BD8F-49D7-9797-758E68B9FC9D}
 		{0F780F3E-8D05-428E-9C78-FEDDDA7DD3D1} = {713ACDCF-301D-4C96-8B75-5BF2A35A326D}
